Script/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
discord.py==2.3.2
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
human-readable==1.3.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
outcome==1.3.0
pycparser==2.21
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
qt-material==2.14
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
random-user-agent==1.0.1
requests==2.31.0
selenium==4.12.0
selenium-stealth==1.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==1.26.18
wsproto==1.2.0
yarl==1.9.2
